requirements.txt
discord-py-interactions
sympy
networkx
# pyjion will require >= .NET 6.0
pyjion # experimenting with JIT